# fl_backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v0/teams": {
      "get": {
        "summary": "List teams, flat or grouped",
        "description": "List teams for a season.\n\n`in_gruppen=true` returns the four groups keyed A-D, otherwise a plain list; check `format` to\ntell them apart. Omitting `saison_id` returns the CURRENT season, and a team not in it is absent.",
        "operationId": "get_teams_api_v0_teams_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "saison_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Saison Id"
            }
          },
          {
            "name": "gruppe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "A",
                    "B",
                    "C",
                    "D"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gruppe"
            }
          },
          {
            "name": "is_disqualified",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Disqualified"
            }
          },
          {
            "name": "in_gruppen",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "In Gruppen"
            }
          },
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          },
          {
            "name": "statistik_scope",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "gruppenphase",
                "gesamt"
              ],
              "type": "string",
              "default": "gruppenphase",
              "title": "Statistik Scope"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1024,
              "minimum": 1,
              "default": 1024,
              "title": "Limit"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "const": "name",
              "type": "string",
              "default": "name",
              "title": "Sort By"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/FLTeamsListResponse"
                    },
                    {
                      "$ref": "#/components/schemas/FLTeamsGroupedResponse"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "format",
                    "mapping": {
                      "list": "#/components/schemas/FLTeamsListResponse",
                      "grouped": "#/components/schemas/FLTeamsGroupedResponse"
                    }
                  },
                  "title": "Response Get Teams Api V0 Teams Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a team",
        "description": "Create a club. `shorthand` is unique across every club, retired ones included, so a duplicate is a 409.",
        "operationId": "post_team_api_v0_teams_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FLPostTeamPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLPostTeamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v0/teams/{team_id}": {
      "get": {
        "summary": "One team",
        "description": "Return one team, with its group and its statistics for a season.\n\nThe path names WHICH TEAM; the query names WHICH SEASON'S figures: `gruppe` and `statistik` are\nseason-scoped. 404 for an id naming no team, and for one with no row that season.",
        "operationId": "get_team_api_v0_teams__team_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Id"
            }
          },
          {
            "name": "saison_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Saison Id"
            }
          },
          {
            "name": "statistik_scope",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "gruppenphase",
                "gesamt"
              ],
              "type": "string",
              "default": "gruppenphase",
              "title": "Statistik Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLTeamsSingleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a team and fan the rename out",
        "description": "Update a club, then rewrite the name and shorthand embedded in its matches.\n\nThe fan-out is not optional: matches carry a copy of both fields, so without it every match card\nshows a stale name indefinitely.",
        "operationId": "patch_team_api_v0_teams__team_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FLPatchTeamPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLPatchTeamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Retire a team (soft delete)",
        "description": "Retire a club from the league. SOFT: `inactive_since` is stamped and the document stays.\n\nRefused while the club is entered in an `active` or `future` season. Its junction rows are left\nalone: the seasons it played still happened.",
        "operationId": "delete_team_api_v0_teams__team_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FLTeamWriteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FLAddress": {
        "properties": {
          "strasse": {
            "type": "string",
            "minLength": 1,
            "title": "Strasse"
          },
          "hausnummer": {
            "type": "string",
            "pattern": "^([0-9\\-abcABC]+)?$",
            "title": "Hausnummer"
          },
          "plz": {
            "type": "string",
            "pattern": "^[0-9]{5}$",
            "title": "Plz"
          },
          "stadtteil": {
            "type": "string",
            "title": "Stadtteil"
          },
          "stadt": {
            "type": "string",
            "minLength": 1,
            "title": "Stadt"
          }
        },
        "type": "object",
        "required": [
          "strasse",
          "hausnummer",
          "plz",
          "stadtteil",
          "stadt"
        ],
        "title": "FLAddress"
      },
...
      "FLDisqualifikation": {
        "properties": {
          "grund": {
            "type": "string",
            "minLength": 1,
            "title": "Grund"
          },
          "datum": {
            "type": "string",
            "pattern": "^[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$",
            "title": "Datum"
          }
        },
        "type": "object",
        "required": [
          "grund",
          "datum"
        ],
        "title": "FLDisqualifikation",
        "description": "Why a team is out of one season, and from when.\n\n`grund` is FREE TEXT and PUBLIC -- it appears on the team's page, and this league publishes no\ndisciplinary code an enum could cite."
      },
      "FLGruppen": {
        "additionalProperties": {
          "items": {
            "$ref": "#/components/schemas/FLTeam"
          },
          "type": "array"
        },
        "propertyNames": {
          "enum": [
            "A",
            "B",
            "C",
            "D"
          ]
        },
        "type": "object",
        "title": "FLGruppen",
        "description": "The four groups, always all four, in standing order.\n\nBuilt by `fl_backend/app/api/teams/services.py :: build_gruppen` alone: the order is the tiebreak\nchain, whose last criterion reads the season's matches."
      },
...
      "FLPatchTeamPayload": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "shorthand": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Shorthand"
          },
          "description": {
            "type": "string",
            "maxLength": 4096,
            "title": "Description"
          },
          "full_name": {
            "type": "string",
            "minLength": 1,
            "title": "Full Name"
          },
          "website_url": {
            "type": "string",
            "title": "Website Url"
          },
          "address": {
            "$ref": "#/components/schemas/FLAddress"
          }
        },
        "type": "object",
        "required": [
          "name",
          "shorthand",
          "description",
          "full_name",
          "website_url",
          "address"
        ],
        "title": "FLPatchTeamPayload"
      },
      "FLPatchTeamResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "updated_document": {
            "$ref": "#/components/schemas/FLTeamRecord"
          },
          "fanned_out_to_spiele": {
            "type": "integer",
            "title": "Fanned Out To Spiele"
          }
        },
        "type": "object",
        "required": [
          "updated_document",
          "fanned_out_to_spiele"
        ],
        "title": "FLPatchTeamResponse"
      },
...
      "FLPostTeamPayload": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "shorthand": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Shorthand"
          },
          "description": {
            "type": "string",
            "maxLength": 4096,
            "title": "Description"
          },
          "full_name": {
            "type": "string",
            "minLength": 1,
            "title": "Full Name"
          },
          "website_url": {
            "type": "string",
            "title": "Website Url"
          },
          "address": {
            "$ref": "#/components/schemas/FLAddress"
          }
        },
        "type": "object",
        "required": [
          "name",
          "shorthand",
          "description",
          "full_name",
          "website_url",
          "address"
        ],
        "title": "FLPostTeamPayload"
      },
      "FLPostTeamResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "created_id": {
            "type": "string",
            "title": "Created Id"
          }
        },
        "type": "object",
        "required": [
          "created_id"
        ],
        "title": "FLPostTeamResponse"
      },
...
      "FLTeam": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "gruppe": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C",
              "D"
            ],
            "title": "Gruppe"
          },
          "statistik": {
            "$ref": "#/components/schemas/FLTeamStatistik"
          },
          "disqualifikation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FLDisqualifikation"
              },
              {
                "type": "null"
              }
            ]
          },
          "shorthand": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Shorthand"
          },
          "description": {
            "type": "string",
            "maxLength": 4096,
            "title": "Description"
          },
          "full_name": {
            "type": "string",
            "minLength": 1,
            "title": "Full Name"
          },
          "website_url": {
            "type": "string",
            "title": "Website Url"
          },
          "address": {
            "$ref": "#/components/schemas/FLAddress"
          },
          "inactive_since": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inactive Since"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "gruppe",
          "statistik",
          "disqualifikation",
          "shorthand",
          "description",
          "full_name",
          "website_url",
          "address",
          "inactive_since"
        ],
        "title": "FLTeam"
      },
...
      "FLTeamRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "shorthand": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Shorthand"
          },
          "description": {
            "type": "string",
            "maxLength": 4096,
            "title": "Description"
          },
          "full_name": {
            "type": "string",
            "minLength": 1,
            "title": "Full Name"
          },
          "website_url": {
            "type": "string",
            "title": "Website Url"
          },
          "address": {
            "$ref": "#/components/schemas/FLAddress"
          },
          "inactive_since": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inactive Since"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "shorthand",
          "description",
          "full_name",
          "website_url",
          "address",
          "inactive_since"
        ],
        "title": "FLTeamRecord",
        "description": "The club document as it is STORED, and what the write endpoints echo.\n\nThe season-scoped fields would mean re-running the team pipeline, whose junction join is strict\n-- and a club being created holds no row yet."
      },
      "FLTeamStatistik": {
        "properties": {
          "anzahl_gespielte_spiele": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Anzahl Gespielte Spiele",
            "default": 0
          },
          "siege": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Siege",
            "default": 0
          },
          "niederlagen": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Niederlagen",
            "default": 0
          },
          "unentschieden": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unentschieden",
            "default": 0
          },
          "tore_geschossen": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tore Geschossen",
            "default": 0
          },
          "tore_kassiert": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tore Kassiert",
            "default": 0
          },
          "punkte": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Punkte",
            "default": 0
          },
          "anzahl_abgesagte_spiele": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Anzahl Abgesagte Spiele",
            "default": 0
          }
        },
        "type": "object",
        "title": "FLTeamStatistik"
      },
...
      "FLTeamWriteResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "updated_document": {
            "$ref": "#/components/schemas/FLTeamRecord"
          }
        },
        "type": "object",
        "required": [
          "updated_document"
        ],
        "title": "FLTeamWriteResponse",
        "description": "What the retire and reactivate endpoints echo. `FLTeamRecord`, for the reason stated on it."
      },
      "FLTeamsGroupedResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "format": {
            "type": "string",
            "const": "grouped",
            "title": "Format",
            "default": "grouped"
          },
          "gruppen": {
            "$ref": "#/components/schemas/FLGruppen"
          },
          "qualifiers_per_group": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Qualifiers Per Group"
          }
        },
        "type": "object",
        "required": [
          "gruppen",
          "qualifiers_per_group"
        ],
        "title": "FLTeamsGroupedResponse",
        "description": "The four groups in standing order, and how many of each advance.\n\n`qualifiers_per_group` rides along rather than being fetched separately, so a page cannot mark a\ncutoff drawn from a different season than the table it marks."
      },
      "FLTeamsListResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "format": {
            "type": "string",
            "const": "list",
            "title": "Format",
            "default": "list"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/FLTeam"
            },
            "type": "array",
            "title": "Teams"
          }
        },
        "type": "object",
        "required": [
          "teams"
        ],
        "title": "FLTeamsListResponse"
      },
...
      "FLTeamsSingleResponse": {
        "properties": {
          "acknowledged": {
            "type": "integer",
            "enum": [
              0,
              1
            ],
            "title": "Acknowledged",
            "default": 1
          },
          "format": {
            "type": "string",
            "const": "single",
            "title": "Format",
            "default": "single"
          },
          "team": {
            "$ref": "#/components/schemas/FLTeam"
          }
        },
        "type": "object",
        "required": [
          "team"
        ],
        "title": "FLTeamsSingleResponse",
        "description": "One team, from `GET /teams/{team_id}`.\n\nNot part of `FLTeamsResponse`: that union discriminates the shapes ONE endpoint can return."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}